feat : 배포로직 추가 및 버그 수정

+++ fitConnect/src/main/java/com/example/fitconnect/service/review/ReviewDeletionService.java
@@ -32,7 +32,9 @@ public class ReviewDeletionService {
 
     @Transactional
     public void deleteReviews(List<Review> reviews) {
-        reviewRepository.deleteAll(reviews);
+        if(reviews != null || !reviews.isEmpty()) {
+            reviewRepository.deleteAll(reviews);
+        }
     }
 }
